(refactor) Move tile styles to styleguide
