+++ packages/esm-commons-lib/src/components/card-summary/summary-card.component.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import styles from './summary-card.scss';
+import styles from '../../styleguide/tiles.scss';
 import { SkeletonText, Tile, Column } from '@carbon/react';
 import { LazyCell } from '../lazy-cell/lazy-cell.component';
 import { OpenmrsEncounter } from '@openmrs/openmrs-form-engine-lib';
