+++ packages/esm-commons-lib/src/index.ts
@@ -47,3 +47,4 @@ export * from './components/expandable-list-widget/expandable-list.component';
 export * from './components/extension-conditional-renderer/patient-based-extension-renderer';
 export * from './components/patient-table/patient-table.component';
 export * from './hooks/useLastEncounter';
+export * from './styleguide/tiles.scss';
